# openapi-facade.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Custom GPT Facade API",
    "version": "2.1.0",
    "description": "BFF (Backend for Frontend) for Custom GPT actions. Provides high-level macros optimized for LLM consumption + unified RPC interface. Timezone: Europe/Prague. All relative time windows (today, lastHour, etc.) are resolved server-side."
  },
  "servers": [
    { "url": "https://api.example.com" }
  ],
...
  "paths": {
    "/macros/inbox/overview": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroInboxOverview",
        "summary": "Lightweight inbox cards (no text)",
        "description": "Get recent emails without snippets. Use timeRange with relative values (today, last24h, thisWeek) for automatic server-side conversion. Max 200 items. Returns subset flag if more available. Ideal for quick overview or drill-down.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InboxOverviewRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email card list with pagination",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InboxOverviewResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/inbox/snippets": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroInboxSnippets",
        "summary": "Inbox with snippets & attachment URLs",
        "description": "Full inbox preview: includes 200-char body snippets + signed URLs for PDF/Excel/CSV files. No server-side categorization—GPT classifies. Attachment URLs expire after 1 hour. Max 50 items per request.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InboxSnippetsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email snippets with attachment metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InboxSnippetsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/email/quickRead": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroEmailQuickRead",
        "summary": "Read 1+ emails with attachments",
        "description": "Fetch full email bodies. Provide ids[] for direct read or searchQuery for search-first. Returns single item or batch array. Always includes attachment metadata + signed download URLs. Format: snippet (brief), minimal (default), full (complete).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailQuickReadRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Single email or batch array",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmailQuickReadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/calendar/plan": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroCalendarPlan",
        "summary": "Daily or weekly schedule view",
        "description": "Get events for a day or week. Computes status (past, ongoing, upcoming) relative to now. Optional: include past events with minimal detail. Useful for 'show my schedule' or 'am I free at X time?'. Max 50 events.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CalendarPlanRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Time range + event list",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CalendarPlanResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/calendar/schedule": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroCalendarSchedule",
        "summary": "Create event (with optional enrichment)",
        "description": "Schedule meeting. If enrichFromContacts=ask and contact exists: returns confirmToken (requires /macros/confirm call). For fixed time: include start/end. For flexibility: provide proposals array (returns first non-conflicting). Conference: none/auto/meet/zoom. Timezone: Europe/Prague.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarScheduleRequest" },
              "examples": {
                "schedule-ask": {
                  "summary": "ASK enrichment flow",
                  "value": {
                    "title": "Meeting with Aleš",
                    "when": {
                      "fixed": {
                        "start": "2025-10-21T14:00:00",
                        "end": "2025-10-21T15:00:00"
                      }
                    },
                    "attendees": [
                      {
                        "email": "ales@example.com",
                        "name": "Aleš"
                      }
                    ],
                    "enrichFromContacts": "ask",
                    "reminders": [15, 60]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event created or confirmation pending",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CalendarScheduleResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/ConflictWithAlternatives" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/calendar/reminderDrafts": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroCalendarReminderDrafts",
        "summary": "Generate reminder emails for today's events",
        "description": "Create email reminders for upcoming meetings. Scans today's events with attendees, generates per-attendee reminder text. Optional template with {title}, {start}, {end}, {location}, {recipientName} placeholders. Can create Gmail drafts (requires manual send) or just return preview.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReminderDraftsRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Draft list with preview + optional draftIds",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReminderDraftsResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/contacts/safeAdd": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroContactsSafeAdd",
        "summary": "Add contacts with deduplication",
        "description": "Bulk add/merge contacts. If dedupeStrategy=ask and duplicates found: returns confirmToken (requires /macros/confirm). Strategy: ask (interactive), merge (auto), keepBoth (both). Uses 'realestate' field for company/org. Max 50 entries per request. Returns created/merged/skipped counts.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ContactsSafeAddRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Contacts created/merged/skipped",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ContactsSafeAddResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/tasks/overview": {
      "post": {
        "tags": ["macros"],
        "operationId": "macroTasksOverview",
        "summary": "Daily/weekly task list",
        "description": "Get tasks grouped by status. Scope: daily (today only), weekly (this week). Optional: include completed. Groups returned: today, thisWeek, overdue, completed. Useful for daily standup or project dashboard.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TasksOverviewRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Tasks grouped by section",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TasksOverviewResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RelativeTime": {
        "type": "string",
        "description": "Relative time window resolved server-side (Europe/Prague TZ). Use instead of manual datetimes for better UX and DST handling.",
        "enum": ["lastHour", "last3h", "last24h", "today", "yesterday", "thisWeek", "last7d"]
      },
      "TimeRange": {
        "oneOf": [
          {
            "type": "object",
            "description": "Relative time window (preferred for GPT queries)",
            "properties": { "relative": { "$ref": "#/components/schemas/RelativeTime" } },
            "required": ["relative"],
            "additionalProperties": false
          },
          {
            "type": "object",
            "description": "Absolute time window (ISO 8601, UTC)",
            "properties": {
              "start": { "type": "string", "format": "date-time", "description": "Inclusive start (ISO 8601)." },
              "end": { "type": "string", "format": "date-time", "description": "Inclusive end (ISO 8601)." }
            },
            "required": ["start", "end"],
            "additionalProperties": false
          }
        ]
      },
...
      "EmailCard": {
        "type": "object",
        "description": "Minimal email card for inbox overview. Lightweight, no body text.",
        "properties": {
          "messageId": { "type": "string" },
          "senderName": { "type": "string" },
          "senderAddress": { "type": "string", "format": "email" },
          "subject": { "type": "string" },
          "receivedAt": { "type": "string", "format": "date-time" },
          "inboxCategory": { "type": "string", "enum": ["primary", "work", "alerts", "newsletters", "social", "updates", "forums", "other"] },
          "label": { "type": ["string", "null"], "description": "Human-readable label name (not Gmail label ID)." }
        },
        "required": ["messageId", "senderAddress", "subject", "receivedAt", "inboxCategory"]
      },
      "EmailSnippetItem": {
        "allOf": [
          { "$ref": "#/components/schemas/EmailCard" },
          {
            "type": "object",
            "properties": {
              "snippet": { "type": "string", "maxLength": 300, "description": "First 200-300 chars of body (auto-truncated)." },
              "attachmentUrls": { "type": "array", "items": { "type": "string" }, "description": "Signed URLs for downloadable attachments (max 50)." }
            }
          }
        ]
      },
      "EmailDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/EmailCard" },
          {
            "type": "object",
            "properties": {
              "body": { "type": ["string", "null"], "description": "Email body (format: snippet/minimal/full)." },
              "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/AttachmentMeta" } }
            }
          }
        ]
      },
      "InboxOverviewRequest": {
        "type": "object",
        "properties": {
          "timeRange": { "$ref": "#/components/schemas/TimeRange" },
          "maxItems": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 },
          "filters": {
            "type": "object",
            "properties": {
              "from": { "type": "string", "description": "Sender email or name fragment." },
              "labelIds": { "type": "array", "items": { "type": "string" }, "maxItems": 10 },
              "hasAttachment": { "type": "boolean" }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "InboxOverviewResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/EmailCard" }, "maxItems": 200 },
          "subset": { "type": "boolean", "description": "True if more results available (pagination needed)." },
          "nextPageToken": { "type": ["string", "null"] }
        },
        "required": ["items", "subset"]
      },
      "InboxSnippetsRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/InboxOverviewRequest" },
          {
            "type": "object",
            "properties": {
              "includeAttachments": { "type": "boolean", "default": true }
            }
          }
        ]
      },
      "InboxSnippetsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/EmailSnippetItem" }, "maxItems": 200 },
          "subset": { "type": "boolean" },
          "nextPageToken": { "type": ["string", "null"] }
        },
        "required": ["items", "subset"]
      },
      "EmailQuickReadRequest": {
        "type": "object",
        "description": "Provide ids[] for direct read or searchQuery for search-first (not both).",
        "properties": {
          "ids": { "type": "array", "items": { "type": "string" }, "minItems": 1, "maxItems": 50 },
          "searchQuery": { "type": "string", "maxLength": 255, "description": "Gmail search syntax (from:, to:, has:attachment, etc.)." },
          "format": { "type": "string", "enum": ["snippet", "minimal", "metadata", "full"], "default": "minimal", "description": "snippet: first 300 chars. minimal: 5KB. full: complete body." }
        },
        "additionalProperties": false,
        "oneOf": [
          { "required": ["ids"] },
          { "required": ["searchQuery"] }
        ]
      },
      "EmailQuickReadResponse": {
        "oneOf": [
          { "type": "object", "properties": { "mode": { "const": "single" }, "item": { "$ref": "#/components/schemas/EmailDetail" } }, "required": ["mode", "item"], "additionalProperties": false },
          { "type": "object", "properties": { "mode": { "const": "batch" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/EmailDetail" }, "maxItems": 50 } }, "required": ["mode", "items"], "additionalProperties": false }
        ]
      },
      "CalendarPlanRequest": {
        "type": "object",
        "properties": {
          "scope": { "type": "string", "enum": ["daily", "weekly"], "description": "daily: single day. weekly: Monday-Sunday." },
          "date": { "type": "string", "format": "date", "description": "Anchor date (YYYY-MM-DD). Resolved in Europe/Prague timezone." },
          "includePast": { "type": "boolean", "default": false },
          "pastTreatment": { "type": "string", "enum": ["hide", "minimal", "full"], "default": "minimal" }
        },
        "required": ["scope", "date"],
        "additionalProperties": false
      },
      "CalendarEvent": {
        "type": "object",
        "properties": {
          "eventId": { "type": "string" },
          "title": { "type": "string" },
          "start": { "type": "string", "format": "date-time" },
          "end": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["past", "ongoing", "upcoming"], "description": "Computed relative to now." },
          "locationText": { "type": ["string", "null"] },
          "mapUrl": { "type": ["string", "null"], "description": "Google Maps link (if location available)." },
          "attendees": { "type": "array", "items": { "type": "object", "properties": { "name": { "type": ["string", "null"] }, "email": { "type": "string", "format": "email" } }, "required": ["email"] } }
        },
        "required": ["eventId", "title", "start", "end", "status"]
      },
      "CalendarPlanResponse": {
        "type": "object",
        "properties": {
          "range": { "type": "object", "properties": { "start": { "type": "string", "format": "date-time" }, "end": { "type": "string", "format": "date-time" }, "tz": { "type": "string" } }, "required": ["start", "end", "tz"] },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CalendarEvent" }, "maxItems": 50 }
        },
        "required": ["range", "items"]
      },
      "CalendarScheduleRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 255 },
          "when": {
            "oneOf": [
              { "type": "object", "properties": { "fixed": { "type": "object", "properties": { "start": { "type": "string", "format": "date-time" }, "end": { "type": "string", "format": "date-time" } }, "required": ["start", "end"], "additionalProperties": false } }, "required": ["fixed"], "additionalProperties": false },
              { "type": "object", "properties": { "proposals": { "type": "array", "items": { "type": "object", "properties": { "start": { "type": "string", "format": "date-time" }, "end": { "type": "string", "format": "date-time" } }, "required": ["start", "end"], "additionalProperties": false }, "minItems": 1, "maxItems": 5 } }, "required": ["proposals"], "additionalProperties": false }
            ]
          },
          "attendees": { "type": "array", "items": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "name": { "type": ["string", "null"] } }, "required": ["email"] }, "maxItems": 20 },
          "enrichFromContacts": { "type": "string", "enum": ["ask", "auto", "off"], "default": "ask", "description": "ask: returns confirmToken if contact found. auto: auto-fill phone/address. off: skip." },
          "conference": { "type": "string", "enum": ["none", "auto", "meet", "zoom"], "default": "none" },
          "location": { "type": ["string", "null"], "maxLength": 500 },
          "notes": { "type": ["string", "null"], "maxLength": 2000 },
          "reminders": { "type": "array", "items": { "type": "integer", "minimum": 1, "maximum": 1440 }, "maxItems": 5, "description": "Minutes before event (15, 30, 60, etc.)." },
          "privacy": { "type": "string", "enum": ["default", "private", "public"], "default": "default" }
        },
        "required": ["title", "when"],
        "additionalProperties": false
      },
      "CalendarScheduleResponse": {
        "oneOf": [
          {
            "type": "object",
            "description": "Event successfully created.",
            "properties": {
              "event": { "$ref": "#/components/schemas/CalendarEvent" },
              "confirmToken": { "type": "null" },
              "warnings": { "type": "array", "items": { "type": "string" } }
            },
            "required": ["event"],
            "additionalProperties": false
          },
          {
            "type": "object",
            "description": "Enrichment opportunity found; awaiting confirmation.",
            "properties": {
              "event": { "type": "null" },
              "confirmToken": { "type": "string", "description": "Pass to /macros/confirm with action=auto-fill or skip." },
              "warnings": { "type": "array", "items": { "type": "string" } }
            },
            "required": ["confirmToken"],
            "additionalProperties": false
          }
        ]
      },
      "ContactsSafeAddRequest": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "maxLength": 255 },
                "email": { "type": ["string", "null"], "format": "email" },
                "phone": { "type": ["string", "null"], "maxLength": 20 },
                "realestate": { "type": ["string", "null"], "maxLength": 500, "description": "Company/org/project name." },
                "notes": { "type": ["string", "null"], "maxLength": 1000 }
              },
              "required": ["name"]
            }
          },
          "dedupeStrategy": { "type": "string", "enum": ["ask", "merge", "keepBoth"], "default": "ask", "description": "ask: returns confirmToken. merge: auto-merge on >70% match. keepBoth: always add new." }
        },
        "required": ["entries"],
        "additionalProperties": false
      },
      "ContactsSafeAddResponse": {
        "type": "object",
        "properties": {
          "created": { "type": "array", "items": { "type": "object", "properties": { "contactId": { "type": "string" }, "name": { "type": "string" }, "email": { "type": "string" } } } },
          "merged": { "type": "array", "items": { "type": "object", "properties": { "into": { "type": "string" }, "from": { "type": "string" }, "mergeScore": { "type": "integer", "minimum": 0, "maximum": 100 } } } },
          "skipped": { "type": "array", "items": { "type": "string" }, "description": "Names/emails that weren't added (duplicates awaiting decision)." },
          "confirmToken": { "type": ["string", "null"], "description": "Present if dedupeStrategy=ask and duplicates found. Call /macros/confirm." }
        },
        "required": ["created", "merged", "skipped"]
      },
      "TasksOverviewRequest": {
        "type": "object",
        "properties": {
          "scope": { "type": "string", "enum": ["daily", "weekly"] },
          "includeCompleted": { "type": "boolean", "default": false },
          "project": { "type": ["string", "null"] },
          "labelIds": { "type": "array", "items": { "type": "string" }, "maxItems": 10 }
        },
        "required": ["scope"],
        "additionalProperties": false
      },
      "TasksOverviewResponse": {
        "type": "object",
        "properties": {
          "sections": {
            "type": "object",
            "properties": {
              "today": { "type": "array", "items": { "type": "object" } },
              "thisWeek": { "type": "array", "items": { "type": "object" } },
              "overdue": { "type": "array", "items": { "type": "object" } },
              "completed": { "type": "array", "items": { "type": "object" } }
            }
          }
        },
        "required": ["sections"]
      },
      "ReminderDraftsRequest": {
        "type": "object",
        "properties": {
          "window": { "type": "string", "enum": ["today", "nextHours"], "description": "today: all events today. nextHours: next N hours (see hours param)." },
          "hours": { "type": "integer", "minimum": 1, "maximum": 24, "description": "Used only when window=nextHours." },
          "template": { "type": ["string", "null"], "maxLength": 1000, "description": "Optional text with {title}, {start}, {end}, {location}, {recipientName} placeholders." },
          "includeLocation": { "type": "boolean", "default": true },
          "createDrafts": { "type": "boolean", "default": true, "description": "If true: creates Gmail drafts (user must send manually). If false: returns preview only." },
          "perAttendee": { "type": "string", "enum": ["separate", "grouped"], "default": "separate" }
        },
        "required": ["window"],
        "additionalProperties": false,
        "allOf": [
          {
            "if": { "properties": { "window": { "const": "nextHours" } } },
            "then": { "required": ["hours"] }
          }
        ]
      },
      "ReminderDraftsResponse": {
        "type": "object",
        "properties": {
          "drafts": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "properties": {
                "draftId": { "type": ["string", "null"], "description": "Gmail draft ID (null if createDrafts=false)." },
                "to": { "type": "string", "format": "email" },
                "subject": { "type": "string" },
                "preview": { "type": "string", "maxLength": 300 },
                "eventId": { "type": "string" }
              },
              "required": ["to", "subject", "preview", "eventId"]
            }
          },
          "subset": { "type": "boolean", "default": false, "description": "True if more drafts available (unlikely)." }
        },
        "required": ["drafts"]
      },
...
    },
    "responses": {
      "ClientError": {
        "description": "4xx client error (invalid request, missing params, enum violation, etc.)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string", "maxLength": 300 },
                "code": { "type": "string", "enum": ["INVALID_PARAM", "BAD_REQUEST", "CONFIRM_SELF_SEND_REQUIRED", "CONFIRMATION_EXPIRED"] },
                "hint": { "type": ["string", "null"], "maxLength": 200 }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 - Not authenticated or token expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
...
      "ConflictWithAlternatives": {
        "description": "409 - Scheduling conflict with alternative suggestions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" },
                "alternatives": { "type": "array", "items": { "type": "object", "properties": { "start": { "type": "string", "format": "date-time" }, "end": { "type": "string", "format": "date-time" } }, "required": ["start", "end"] }, "maxItems": 5 }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
...
      "ServerError": {
        "description": "5xx server error (temporary issue, please retry)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "RateLimit429": {
        "description": "Too Many Requests - rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "rate_limit_exceeded" },
                "retryAfterSeconds": { "type": "integer", "minimum": 1 }
              },
              "required": ["error"]
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": { "type": "integer", "minimum": 1 }
          }
        }
      }
    }
  }
}